tambah sukses

+++ database/migrations/2021_03_02_035015_create_barang_table.php
@@ -18,7 +18,7 @@ class CreateBarangTable extends Migration
             $table->string('NamaBarang');
             $table->integer('Jumlah');
             $table->integer('Harga');
-            $table->integer('Diskon');
+            $table->float('Diskon');
             $table->date('Kadaluarsa');
             $table->timestamps();
         });
